reworked modular reduction, public key in NTT domain

+++ ref/cbd.c
@@ -1,17 +1,16 @@
-#include <stdlib.h>
-
+#include <stdint.h>
+#include "params.h"
 #include "cbd.h"
 
 /*************************************************
-* Name:        load_littleendian
+* Name:        load32_littleendian
 *
-* Description: load bytes into a 64-bit integer
+* Description: load bytes into a 32-bit integer
 *              in little-endian order
 *
 * Arguments:   - const unsigned char *x: pointer to input byte array
-*              - bytes:                  number of bytes to load, has to be <= 8
 *
-* Returns 64-bit unsigned integer loaded from x
+* Returns 32-bit unsigned integer loaded from x
 **************************************************/
 static uint32_t load32_littleendian(const unsigned char *x)
 {
@@ -36,14 +35,14 @@ static uint32_t load32_littleendian(const unsigned char *x)
 void cbd(poly *r, const unsigned char *buf)
 {
 #if KYBER_ETA == 2
-  uint32_t d, t;
+  uint32_t d,t;
   int16_t a,b;
-  size_t i,j;
+  int i,j;
 
   for(i=0;i<KYBER_N/8;i++)
   {
     t = load32_littleendian(buf+4*i);
-    d = t & 0x55555555;
+    d  = t & 0x55555555;
     d += (t>>1) & 0x55555555;
 
     for(j=0;j<8;j++)
